// src/app/core/model/list-column.model.ts
export class ListColumn {
  name?: string;
  property?: string;
  visible?: boolean;
  searchable?: boolean;
  searchValue?: any;
  isModelProperty?: boolean;
  displayFn: any;
  datatype?: any;
  searchfield?: any;
  sortOrder?: any;
  searchPlaceholder?: any;
  filter_type?: any;
  column_width?: any;
}